refactor(site): move DynamicClientRegistrationSetting to its own file
